requirements.txt
fastapi
uvicorn
requests
python-dotenv
apscheduler
openai
feedparser
jinja2
